requirements.txt
Django==3.2.5
djangorestframework==3.12.4
entrypoints==0.3
idna==2.10
ipykernel==5.5.3
ipython==7.23.0
joblib==1.0.1
matplotlib-inline==0.1.2
numpy==1.20.3
pandas==1.2.4
scikit-learn==0.24.2
scipy==1.6.3
sklearn==0.0
xgboost==1.4.2
gunicorn
django-heroku